Splitting out akka-http and http4s sample projects
