+++ .gitignore
@@ -1,11 +1,11 @@
 .DS_Store
 target/
 
-modules/sample/target/
-modules/sample/src/main/scala/*
-modules/sample/src/main/scala/generated/*
-!modules/sample/src/main/scala/App.scala
-!modules/sample/src/main/scala/support
-!modules/sample/src/test/resources/
+modules/sample*/target/
+modules/sample*/src/main/scala/*
+modules/sample*/src/main/scala/generated/*
+!modules/sample*/src/main/scala/App.scala
+!modules/sample*/src/main/scala/support
+!modules/sample*/src/test/resources/
 
 .idea
