+++ modules/sample-http4s/src/test/scala/core/Http4s/Http4sFullTracerTest.scala
